Add a global data storage
This is used to make the tracked data survive disable/enable cycles, making it
persistent for the whole session.

The problem here was that when the screen is locked, extensions are disabled
and, thus, the tracked data is reset. In this way tracked data used to generate
statistics is preserved until the user actually logs out.